New: Exercise List integration
