+++ exercises/templates/exercises/follow_line_react/exercise.html
@@ -1,5 +1,4 @@
 {% extends "react_frontend/exercise_base.html" %}
 
-{% block content %}
-    HERE COMES THE CONTENT
+{% block react-content %}
 {% endblock %}
