+++ react_frontend/src/components/DrawerAppBar.js
@@ -1,0 +1,89 @@
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Divider from '@mui/material/Divider';
+import Drawer from '@mui/material/Drawer';
+import IconButton from '@mui/material/IconButton';
+import List from '@mui/material/List';
+import ListItemButton from '@mui/material/ListItemButton';
+import MenuIcon from '@mui/icons-material/Menu';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import Button from '@mui/material/Button';
+import ForumIcon from '@mui/icons-material/Forum';
+
+const drawerWidth = 240;
+const navItems = ['forum'];
+
+const DrawerAppBar = (props) => {
+    const { window } = props;
+    const [mobileOpen, setMobileOpen] = React.useState(false);
+
+    const handleDrawerToggle = () => {
+        setMobileOpen(!mobileOpen);
+    };
+
+    const drawer = (
+        <Box onClick={handleDrawerToggle} sx={{ textAlign: 'center' }}>
+            <Typography variant="h6" sx={{ my: 2 }}>
+                Robotics Academy by JdeRobot
+            </Typography>
+            <Divider />
+            <List>
+
+                        <ListItemButton sx={{ textAlign: 'center' }} onClick={location.href='https://forum.jderobot.org/c/english/roboticsacademy/12'}>
+                            Forum
+                        </ListItemButton>
+            </List>
+        </Box>
+    );
+
+    const container = window !== undefined ? () => window().document.body : undefined;
+
+    return (
+        <Box sx={{ display: 'flex'}} mb={10}>
+            <AppBar component="nav">
+                <Toolbar>
+                    <IconButton
+                        color="inherit"
+                        aria-label="open drawer"
+                        edge="start"
+                        onClick={handleDrawerToggle}
+                        sx={{ mr: 2, display: { sm: 'none' } }}
+                    >
+                        <MenuIcon />
+                    </IconButton>
+                    <Typography
+                        variant="h6"
+                        component="div"
+                        sx={{ flexGrow: 1, display: { xs: 'none', sm: 'block' } }}
+                    >
+                        Robotics Academy by JdeRobot
+                    </Typography>
+                    <Button variant="outlined" sx={{ display: { xs: 'none', sm: 'block'}, color: '#fff' }} startIcon={<ForumIcon />} onClick={location.href='https://forum.jderobot.org/c/english/roboticsacademy/12'}>
+                        Forum
+                    </Button>
+                </Toolbar>
+            </AppBar>
+            <Box component="nav">
+                <Drawer
+                    container={container}
+                    variant="temporary"
+                    open={mobileOpen}
+                    onClose={handleDrawerToggle}
+                    ModalProps={{
+                        keepMounted: true, // Better open performance on mobile.
+                    }}
+                    sx={{
+                        display: { xs: 'block', sm: 'none' },
+                        '& .MuiDrawer-paper': { boxSizing: 'border-box', width: drawerWidth },
+                    }}
+                >
+                    {drawer}
+                </Drawer>
+            </Box>
+        </Box>
+    );
+}
+
+export default DrawerAppBar;
