+++ react_frontend/src/components/exercises/AceEditor.js
@@ -1,0 +1,36 @@
+import editor from "react-ace";
+import * as React from 'react';
+import {Box} from "@mui/material";
+import { VncScreen } from 'react-vnc';
+
+function AceEditor() {
+
+    return (
+        <Box sx={{m:3}}>
+{/*            <AceEditor*/}
+{/*                mode="python"*/}
+{/*                theme="github"*/}
+{/*                name="blah2"*/}
+{/*                fontSize={14}*/}
+{/*                showPrintMargin={true}*/}
+{/*                showGutter={true}*/}
+{/*                highlightActiveLine={true}*/}
+{/*                value={`from GUI import GUI*/}
+{/*from HAL import HAL*/}
+{/*# Enter sequential code!*/}
+
+{/*while True:*/}
+{/*    # Enter iterative code!`}*/}
+{/*                setOptions={{*/}
+{/*                    enableBasicAutocompletion: true,*/}
+{/*                    enableLiveAutocompletion: true,*/}
+{/*                    enableSnippets: true,*/}
+{/*                    showLineNumbers: true,*/}
+{/*                    tabSize: 2,*/}
+{/*                }}/>*/}
+
+        </Box>
+    );
+}
+
+export default AceEditor;
