+++ react_frontend/templates/base.html
@@ -20,8 +20,6 @@
     <!--CSS Main -->
     <link rel="stylesheet" type="text/css" href="{% static 'exercises/assets/css/main.css' %}">
 
-    {% render_bundle 'main' 'css' %}
-
     {% block header %}
         <!-- Here comes additions to header section -->
     {%  endblock %}
